Translate from German and update task beep instead.
Since I'm english, I've updated the message prompts to english and
I've replaced playing an audio file with a call to [Console]::Beep.

The frequency is 24kHz for a duration of 100ms. Just long enough
to keep sound bar active. Motivation being I don't want to hear
a sound being played just to fix a limitation on an otherwise good
devise.